@@ -1,5 +1,10 @@
 package jp.kshoji.blehid.sample;
 
+import static android.view.MotionEvent.ACTION_DOWN;
+import static android.view.MotionEvent.ACTION_POINTER_DOWN;
+import static android.view.MotionEvent.ACTION_POINTER_UP;
+import static android.view.MotionEvent.ACTION_UP;
+
 import android.content.Context;
 import android.hardware.Sensor;
 import android.hardware.SensorEvent;
@@ -10,19 +15,13 @@ import android.view.MotionEvent;
 import android.view.View;
 import android.view.View.OnTouchListener;
 
-import jp.kshoji.blehid.JoystickPeripheral;
-import jp.kshoji.blehid.sample.R.id;
-import jp.kshoji.blehid.sample.R.layout;
-import jp.kshoji.blehid.sample.R.string;
+import com.zero.ble_hid.R;
 
-import static android.view.MotionEvent.ACTION_DOWN;
-import static android.view.MotionEvent.ACTION_POINTER_DOWN;
-import static android.view.MotionEvent.ACTION_POINTER_UP;
-import static android.view.MotionEvent.ACTION_UP;
+import jp.kshoji.blehid.JoystickPeripheral;
 
 /**
  * Activity for BLE Joystick peripheral
- * 
+ *
  * @author K.Shoji
  */
 public class JoystickActivity extends AbstractBleActivity implements SensorEventListener {
@@ -32,19 +31,19 @@ public class JoystickActivity extends AbstractBleActivity implements SensorEvent
 
     private SensorManager sensorManager;
     private Sensor accelerometerSensor;
-    
+
     @Override
     protected void onCreate(final Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(layout.activity_joystick);
+        setContentView(R.layout.activity_joystick);
 
-        setTitle(getString(string.ble_joystick));
-        
+        setTitle(getString(R.string.ble_joystick));
+
         sensorManager = (SensorManager) getSystemService(Context.SENSOR_SERVICE);
         accelerometerSensor = sensorManager.getDefaultSensor(Sensor.TYPE_ACCELEROMETER);
         sensorManager.registerListener(this, accelerometerSensor, SensorManager.SENSOR_DELAY_UI);
-        
-        findViewById(id.leftButton).setOnTouchListener(new OnTouchListener() {
+
+        findViewById(R.id.leftButton).setOnTouchListener(new OnTouchListener() {
             @Override
             public boolean onTouch(final View view, final MotionEvent motionEvent) {
                 switch (motionEvent.getAction()) {
@@ -61,7 +60,7 @@ public class JoystickActivity extends AbstractBleActivity implements SensorEvent
                 return false;
             }
         });
-        findViewById(id.middleButton).setOnTouchListener(new OnTouchListener() {
+        findViewById(R.id.middleButton).setOnTouchListener(new OnTouchListener() {
             @Override
             public boolean onTouch(final View view, final MotionEvent motionEvent) {
                 switch (motionEvent.getAction()) {
@@ -78,7 +77,7 @@ public class JoystickActivity extends AbstractBleActivity implements SensorEvent
                 return false;
             }
         });
-        findViewById(id.rightButton).setOnTouchListener(new OnTouchListener() {
+        findViewById(R.id.rightButton).setOnTouchListener(new OnTouchListener() {
             @Override
             public boolean onTouch(final View view, final MotionEvent motionEvent) {
                 switch (motionEvent.getAction()) {
@@ -100,15 +99,16 @@ public class JoystickActivity extends AbstractBleActivity implements SensorEvent
     @Override
     void setupBlePeripheralProvider() {
         joystick = new JoystickPeripheral(this);
-        joystick.setDeviceName(getString(string.ble_joystick));
+        joystick.setDeviceName(getString(R.string.ble_joystick));
         joystick.startAdvertising();
     }
 
     private final float[] gravity = new float[3];
     private final float[] linear_acceleration = new float[3];
     private final float[] velocity = new float[3];
+
     @Override
-    public void onSensorChanged(final SensorEvent event){
+    public void onSensorChanged(final SensorEvent event) {
         final float alpha = 0.8f;
 
         gravity[0] = alpha * gravity[0] + (1 - alpha) * event.values[0];
@@ -118,11 +118,11 @@ public class JoystickActivity extends AbstractBleActivity implements SensorEvent
         linear_acceleration[0] = event.values[0] - gravity[0];
         linear_acceleration[1] = event.values[1] - gravity[1];
         linear_acceleration[2] = event.values[2] - gravity[2];
-        
+
         velocity[0] += linear_acceleration[0];
         velocity[1] += linear_acceleration[1];
         velocity[2] += linear_acceleration[2];
-        
+
         if (joystick != null) {
             joystick.movePointer((int) velocity[0], (int) velocity[1], (int) velocity[2], left, middle, right);
         }
@@ -137,7 +137,7 @@ public class JoystickActivity extends AbstractBleActivity implements SensorEvent
     protected void onDestroy() {
         super.onDestroy();
         sensorManager.unregisterListener(this, accelerometerSensor);
-                
+
         if (joystick != null) {
             joystick.stopAdvertising();
         }
