@@ -5,14 +5,13 @@ import android.view.View;
 import android.view.View.OnClickListener;
 import android.widget.TextView;
 
+import com.zero.ble_hid.R;
+
 import jp.kshoji.blehid.KeyboardPeripheral;
-import jp.kshoji.blehid.sample.R.id;
-import jp.kshoji.blehid.sample.R.layout;
-import jp.kshoji.blehid.sample.R.string;
 
 /**
  * Activity for BLE Keyboard peripheral
- * 
+ *
  * @author K.Shoji
  */
 public class KeyboardActivity extends AbstractBleActivity {
@@ -22,15 +21,15 @@ public class KeyboardActivity extends AbstractBleActivity {
     @Override
     protected void onCreate(final Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(layout.activity_keyboard);
+        setContentView(R.layout.activity_keyboard);
 
-        setTitle(getString(string.ble_keyboard));
+        setTitle(getString(R.string.ble_keyboard));
 
-        findViewById(id.sendButton).setOnClickListener(new OnClickListener() {
+        findViewById(R.id.sendButton).setOnClickListener(new OnClickListener() {
             @Override
             public void onClick(final View view) {
                 if (keyboard != null) {
-                    keyboard.sendKeys(((TextView) findViewById(id.editText)).getText().toString());
+                    keyboard.sendKeys(((TextView) findViewById(R.id.editText)).getText().toString());
                 }
             }
         });
@@ -39,14 +38,14 @@ public class KeyboardActivity extends AbstractBleActivity {
     @Override
     void setupBlePeripheralProvider() {
         keyboard = new KeyboardPeripheral(this);
-        keyboard.setDeviceName(getString(string.ble_keyboard));
+        keyboard.setDeviceName(getString(R.string.ble_keyboard));
         keyboard.startAdvertising();
     }
 
     @Override
     protected void onDestroy() {
         super.onDestroy();
-        
+
         if (keyboard != null) {
             keyboard.stopAdvertising();
         }
