@@ -2,46 +2,46 @@ package jp.kshoji.blehid.sample;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.view.View.OnClickListener;
 
-import jp.kshoji.blehid.sample.R.id;
-import jp.kshoji.blehid.sample.R.layout;
-import jp.kshoji.blehid.sample.R.string;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.zero.ble_hid.R;
+
 
 /**
  * Main Activity
- * 
+ *
  * @author K.Shoji
  */
 public class MainActivity extends AppCompatActivity {
     @Override
     protected void onCreate(final Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(layout.activity_main);
+        setContentView(R.layout.activity_main);
 
-        setTitle(getString(string.ble_hid));
-        
-        findViewById(id.mouseButton).setOnClickListener(new OnClickListener() {
+        setTitle(getString(R.string.ble_hid));
+
+        findViewById(R.id.mouseButton).setOnClickListener(new OnClickListener() {
             @Override
             public void onClick(View view) {
                 startActivity(new Intent(getApplicationContext(), MouseActivity.class));
             }
         });
-        findViewById(id.absoluteMouseButton).setOnClickListener(new OnClickListener() {
+        findViewById(R.id.absoluteMouseButton).setOnClickListener(new OnClickListener() {
             @Override
             public void onClick(View view) {
                 startActivity(new Intent(getApplicationContext(), AbsoluteMouseActivity.class));
             }
         });
-        findViewById(id.keyboardButton).setOnClickListener(new OnClickListener() {
+        findViewById(R.id.keyboardButton).setOnClickListener(new OnClickListener() {
             @Override
             public void onClick(View view) {
                 startActivity(new Intent(getApplicationContext(), KeyboardActivity.class));
             }
         });
-        findViewById(id.joystickButton).setOnClickListener(new OnClickListener() {
+        findViewById(R.id.joystickButton).setOnClickListener(new OnClickListener() {
             @Override
             public void onClick(View view) {
                 startActivity(new Intent(getApplicationContext(), JoystickActivity.class));
